+++ src/test/java/com/keboola/tableexporter/TableDataTest.java
@@ -48,7 +48,7 @@ public class TableDataTest extends BaseTest {
                 when(resultSetMock.getString("TABLESPACE_NAME")).thenReturn("sometablespace");
                 when(resultSetMock.getString("OWNER")).thenReturn("something");
                 when(resultSetMock.getString("NUM_ROWS")).thenReturn("127");
-                when(resultSetMock.getInt("NUM_ROWS")).thenReturn(127);
+                when(resultSetMock.getLong("NUM_ROWS")).thenReturn(127L);
                 break;
             case "noRowCountTest":
                 when(resultSetMock.getString("TABLE_NAME")).thenReturn("someTableName");
